features for getting inscrits from android mobile app with post request and 'nfc_code' in DB

+++ downloadInscrits.php
@@ -7,18 +7,29 @@ GLOBAL $CONFIG_MDP_ORGA;
 
 $authen_orga = valider('authentifie_orga', 'SESSION');
 $mdp_orga = valider('mdp', 'POST');
+$output_type = valider('type', 'POST');
+
 if($authen_orga || ($mdp_orga && $mdp_orga === $CONFIG_MDP_ORGA)){
-	$key = generate_res_key();
+	if ($output_type == 'json') {
 
-	$files = scandir("res");
+		$data = array();
+		$data["success"] = true;
+		$data["result"] = sql_select_inscrits();
 
-	for($i=2; $i<count($files); $i++){
-		unlink("res/".$files[$i]);
+		echo json_encode($data);
+	} else {
+	
+		$key = generate_res_key();
+		$files = scandir("res");
+	
+		for($i=2; $i<count($files); $i++){
+			unlink("res/".$files[$i]);
+		}
+	
+		$filename = "res/temp_inscrits_" . $key . ".csv";
+		sql_select_inscrits_csv($filename);
+		header("Location: $filename");
 	}
-
-	$filename = "res/temp_inscrits_" . $key . ".csv";
-	sql_select_inscrits_csv($filename);
-	header("Location: $filename");
 }
 else{
 	include('templates/ca_id/orga_id_cheat.php');
